requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
python-dotenv==1.0.1
google-generativeai==0.3.2
openai==1.12.0
discord.py==2.3.2
pyyaml==6.0.1
lxml==5.1.0
